Changed several AVMPLUS_AMD64 ifdefs to AVMPLUS_64BIT (r=treilly+)
Several places in MMgc, and a few elsewhere, use the wrong ifdef switches.
PPC64 is coming down the pike which requires being exact about AMD64 vs 64BIT
in general.

+++ MMgc/GCAlloc.cpp
@@ -453,7 +453,7 @@ start:
 					if (marks & kFinalize)
 					{     
 						GCFinalizable *obj = (GCFinalizedObject*)GetUserPointer(item);
-						GCAssert(*(int*)obj != 0);
+						GCAssert(*(intptr_t*)obj != 0);
 						obj->~GCFinalizable();
 						bits[i] &= ~(kFinalize<<(j*4));
